+++ CHANGELOG.md
@@ -60,6 +60,7 @@
 ## 2.3.1 (2020-11-23)
 
 - Improved column names for uploads
+- Allow to configure time range selector preset ranges
 
 ## 2.3.0 (2020-11-16)